Add GitLabCI source
This commit adds GitLabCI as a CI source. Merge request ID isn't
available as ENV variable with GitLab CI runners so I'm using project ID
to fetch list of merge requests for given project, and branch name to
find specific merge request that triggered the build. I'm not sure if
this is the right way to do it, and what happens if build is not
triggered with merge request.